# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Librarie API",
        "version": "0.1"
    },
    "paths": {
...
        "/user": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update email-ul user-ului autentificat",
                "description": "Update email-ul user-ului autentificat",
                "operationId": "c1f4bf30328d3929c1d0563d017d5710",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Adauga clientul",
                "description": "Adauga clientul",
                "operationId": "ce3cf31668a69eb39e5e98738715afa1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nume": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Petrache"
                                    },
                                    "prenume": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Cristina"
                                    },
                                    "email": {
                                        "description": "",
                                        "type": "string",
                                        "example": "cristina.petrache@hotmail.com"
                                    },
                                    "parola": {
                                        "description": "",
                                        "type": "string",
                                        "example": "asd123"
                                    },
                                    "nr_telefon": {
                                        "description": "",
                                        "type": "string",
                                        "example": "0755158838"
                                    },
                                    "adresa": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Str. Padurii, Nr. 85"
                                    },
                                    "oras": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Sabareni"
                                    },
                                    "judet": {
                                        "description": "",
                                        "type": "string",
                                        "example": "Giurgiu"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/user/parola": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update parola user-ului autentificat",
                "description": "Update parola user-ului autentificat",
                "operationId": "dfb204f027786e31d4072e6d3908e170",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parola": {
                                        "description": "",
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}